+++ kernel/fs/ext2/ext2_fs.c
@@ -63,7 +63,7 @@ static size_t position(void *data)
   return edata->offset;
 }
 
-static int resize(void *data, size_t size)
+static int resize(void *data, uint64_t size)
 {
   ext2_vfs_data_t *edata = data;
   return ext2_inode_iterator_resize(edata->it, size);
